+++ lab-4-rocketmq/lab-4-rocketmq-demo/src/main/java/ustc/sse/springboot/lab04/consumer/Demo02Consumer.java
@@ -1,0 +1,28 @@
+package ustc.sse.springboot.lab04.consumer;
+
+import org.apache.rocketmq.spring.annotation.RocketMQMessageListener;
+import org.apache.rocketmq.spring.core.RocketMQListener;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+import ustc.sse.springboot.lab04.message.Demo02Message;
+
+/**
+ * @author MonoSirius
+ * @date 2023/6/10
+ */
+@Component
+@RocketMQMessageListener(
+        topic = Demo02Message.TOPIC,
+        consumerGroup = "demo02-consumer-group-" + Demo02Message.TOPIC
+)
+public class Demo02Consumer implements RocketMQListener<Demo02Message> {
+
+    private Logger logger = LoggerFactory.getLogger(getClass());
+    @Override
+    public void onMessage(Demo02Message message) {
+        logger.info("[onMessage][线程编号:{} 消息内容：{}]", Thread.currentThread().getId(), message);
+    }
+
+}
+